db-pipeline/requirements.txt
numpy==2.0.1
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
tzdata==2024.1